Advent2020: created base classes & script for re-use
Using Day #2 as an example, this was an attempt to try and create some
base classes and a common run.sh script that could be re-used throughout
the code competition.  There seemed to be a common pattern in terms of
how the input files were being parsed so the base classes try to codify
that.  The run.sh script simply tries to make it easy to run and verify
each day's code against its inputs and answers.